requirements.txt
fastapi
uvicorn
openai
python-dotenv
python-docx
pydantic
supabase